Implement init function
This implements the init function for the class. It is already capable
of working with nd-arrays. This is because if I plan to do it for
higher dimensional arrays anyways there is no reason to rewrite a bunch
of code because I made the original init only work for 1d-arrays. It
uses a recursive fill_list function to create the necessary lists.
Error checking and documentation is minimal.