Change the way errors are handled
It doesn't make sense for array_class to handle errors. It should simply
raise them and it is up to the caller to handle them. For example, if
you call "np.ones((2,2)) + np.ones((3,3))" numpy doesnt handle this, it
raises ValueError and you have to handle it. Thereforevery error is now
raised instead of simply printed to console.